# Use bump allocator for keys