--- a/rust/optimized/Cargo.toml
+++ b/rust/optimized/Cargo.toml
@@ -9,6 +9,7 @@ name = "countwords"
 path = "main.rs"
 
 [dependencies]
+bumpalo = "3.6.1"
 fxhash = "0.2.1"
 
 [profile.release]