Should not reallocate on every keys

Before

```
> hyperfine 'target/release/countwords < ../../kjvbible.txt'
Benchmark #1: target/release/countwords < ../../kjvbible.txt
  Time (mean ± σ):      60.7 ms ±  11.8 ms    [User: 53.3 ms, System: 7.5 ms]
  Range (min … max):    43.9 ms …  74.8 ms    44 runs
```

After keys arena

```
> hyperfine 'target/release/countwords < ../../kjvbible.txt'
Benchmark #1: target/release/countwords < ../../kjvbible.txt
  Time (mean ± σ):      47.7 ms ±  11.5 ms    [User: 41.2 ms, System: 6.5 ms]
  Range (min … max):    37.4 ms …  70.0 ms    48 runs
 ```

After bufwriter

```
> hyperfine 'target/release/countwords < ../../kjvbible.txt'
Benchmark #1: target/release/countwords < ../../kjvbible.txt
  Time (mean ± σ):      46.7 ms ±  13.5 ms    [User: 43.4 ms, System: 3.3 ms]
  Range (min … max):    29.9 ms …  63.9 ms    43 runs
```

Follow up of https://github.com/benhoyt/countwords/pull/1#issuecomment-799584798
cc @BurntSushi