@pickfire 2021-03-17
By the way, I used buffered writer for standard output. So it became faster again slightly. 60ms -> 50ms -> 45ms

The dependencies is just
```
├── bumpalo v3.6.1
└── fxhash v0.2.1
    └── byteorder v1.4.3
```
Still need to see whether @benhoyt allows this, bump allocator is not available in the standard library and probably cannot be written without unsafe, but it makes sense to use it here since we shouldn't allocate on every new key insert. Still, using a bump allocator is kinda cheating in a way since that will allocate a huge buffer, well GC languages does that too. So fair right?

If we want to cheat we can just do it at constant time, like https://blog.mgattozzi.dev/oxidizing-the-technical-interview/