@BurntSushi 2021-03-17
Ah yeah, I see, using a `Vec<u8>` is problematic because the code wants to borrow it in the hashmap and mutably at the same time.

Dunno. I could go either way.  Like, yeah, maybe that's what GC languages do. But then again, that's their advantage. It's hard to draw the line. That's probably why it was just originally "no third party libs."

> or use ahead of time HashMap capacity allocation, but we didn't.

I tried that but couldn't see a difference. But that's a fair optimization if it matters I think.