@benhoyt 2021-03-17
Yeah, I'd like to stick as closely to the stdlib as possible, given the constraints I set. I think it's useful to show what a language can do "out of the box". Like you say, you *could* just copy the bump allocator code in here, though it looks like a reasonable amount of code, and I don't really think this is in the spirit of the constraints anymore. So closing this one. Thanks anyway!