@BurntSushi 2021-03-18
Yeah, I've know about bumpalo. Just never used it or happened to stumble across it solving a problem so elegantly in isolation like this.